requirements.txt
python-dotenv==1.0.1;
langchain==0.3.3;
langchain-openai==0.2.2;
chromadb==0.5.23;
langchain-chroma==0.2.0;
rank_bm25==0.2.2;
flashrank==0.2.10;
langchain-community==0.3.2;
langgraph==0.2.62;
typing_extensions==4.12.2;
tqdm==4.66.4;
pypdf==5.7.0;
chainlit==2.5.5;